switch to double, added builtins and parse doubles

+++ src/main/java/icl/ast/AstNum.java
@@ -1,9 +1,9 @@
 package icl.ast;
 
 public class AstNum extends AstNode {
-	public final float value;
+	public final double value;
 
-	public AstNum(short value) {
+	public AstNum(double value) {
 		this.value = value;
 	}
